校验逻辑修改

+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/impl/WorkStationServiceImpl.java
@@ -17,6 +17,7 @@ import com.jd.bluedragon.distribution.station.domain.WorkStation;
 import com.jd.bluedragon.distribution.station.domain.WorkStationCountVo;
 import com.jd.bluedragon.distribution.station.query.WorkStationQuery;
 import com.jd.bluedragon.distribution.station.service.WorkStationService;
+import com.jd.bluedragon.distribution.utils.CheckHelper;
 import com.jd.bluedragon.dms.utils.DmsConstants;
 import com.jd.bluedragon.utils.StringHelper;
 import com.jd.ql.dms.common.web.mvc.api.PageDto;
@@ -46,7 +47,10 @@ public class WorkStationServiceImpl implements WorkStationService {
 	 * @return
 	 */
 	public Result<Boolean> insert(WorkStation insertData){
-		Result<Boolean> result = Result.success();
+		Result<Boolean> result = checkAndFillBeforeAdd(insertData);
+		if(!result.isSuccess()) {
+			return result;
+		}
 		insertData.setBusinessKey(DmsConstants.CODE_PREFIX_WORK_STATION.concat(StringHelper.padZero(this.genObjectId.getObjectId(WorkStation.class.getName()),11)));
 		result.setData(workStationDao.insert(insertData) == 1);
 		return result;
@@ -74,6 +78,21 @@ public class WorkStationServiceImpl implements WorkStationService {
 		return result;
 	}
 	/**
+	 * 校验并填充数据add
+	 * @param insertData
+	 * @return
+	 */
+	private Result<Boolean> checkAndFillBeforeAdd(WorkStation insertData){
+		Result<Boolean> result = checkAndFillNewData(insertData);
+		if(!result.isSuccess()) {
+			return result;
+		}
+		if(this.isExist(insertData)) {
+			return result.toFail("该作业区、工序已存在，请修改！");
+		}
+		return result;
+	}	
+	/**
 	 * 校验并填充导入数据
 	 * @param dataList
 	 * @return
@@ -87,25 +106,50 @@ public class WorkStationServiceImpl implements WorkStationService {
 		int rowNum = 1;
 		for(WorkStation data : dataList) {
 			String rowKey = "第" + rowNum + "行";
-			String workCode = data.getWorkCode();
-			String areaCode = data.getAreaCode();
-			if(StringHelper.isEmpty(workCode)) {
-				return result.toFail(rowKey + "工序编码为空！");
-			}
-			if(StringHelper.isEmpty(areaCode)) {
-				return result.toFail(rowKey + "作业区编码为空！");
+			Result<Boolean> result0 = checkAndFillNewData(data);
+			if(!result0.isSuccess()) {
+				return result0.toFail(rowKey + result0.getMessage());
 			}
 			rowNum ++;
 		}
 		return result;
-	}	
+	}
+	/**
+	 * 校验并填充数据
+	 * @param data
+	 * @return
+	 */
+	private Result<Boolean> checkAndFillNewData(WorkStation data){
+		Result<Boolean> result = Result.success();
+		String workCode = data.getWorkCode();
+		String workName = data.getWorkName();
+		String areaCode = data.getAreaCode();
+		String areaName = data.getAreaName();
+		
+		if(!CheckHelper.checkStr("作业区ID", areaCode, 50, result).isSuccess()) {
+			return result;
+		}		
+		if(!CheckHelper.checkStr("作业区名称", areaName, 100, result).isSuccess()) {
+			return result;
+		}
+		if(!CheckHelper.checkStr("工序ID", workCode, 50, result).isSuccess()) {
+			return result;
+		}
+		if(!CheckHelper.checkStr("工序名称", workName, 100, result).isSuccess()) {
+			return result;
+		}
+		return result;
+	}
 	/**
 	 * 根据id更新数据
 	 * @param updateData
 	 * @return
 	 */
 	public Result<Boolean> updateById(WorkStation updateData){
-		Result<Boolean> result = Result.success();
+		Result<Boolean> result = checkAndFillNewData(updateData);
+		if(!result.isSuccess()) {
+			return result;
+		}
 		workStationDao.deleteById(updateData);
 		updateData.setId(null);
 		result.setData(workStationDao.insert(updateData) == 1);
